OceanMonitoringSystem/PythonAnalysisService/requirements.txt
grpcio>=1.50.0
grpcio-tools>=1.50.0
numpy>=1.24.0
